backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
SQLAlchemy==2.0.25
python-dotenv==1.0.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
email-validator==2.1.0.post1
pydantic==1.10.13
httpx==0.27.0
svix==1.19.0
